made several visual changes to the ui
