+++ src/_colorPalette.scss
@@ -2,6 +2,7 @@ $themes: (
   light: (
     primary: white,
     secondary: black,
+    text-shad: 1px 1px white,
     accent: #0000ff33,
     accent-2: #9500ff,
     background-1: #cccdfb,
@@ -10,10 +11,15 @@ $themes: (
     border-1: 2px solid black,
     underline: 1px solid black,
     mochi-guy: url("./assets/mochiSpinCool3.png"),
+    btn-grad: linear-gradient(160deg, rgba(228, 253, 0, 0.762) 0%, rgba(253, 207, 0, 0.76) 100%),
+    btn-grad-hover: linear-gradient(160deg, rgba(194, 185, 63, 0.762) 0%, rgba(194, 185, 63, 0.76) 100%),
+    bg-opacity: 0.9,
+    bg-image-op: rgba(255,255,255, 0.9),
   ),
   dark: (
     primary: #2a2a2e,
     secondary: white,
+    text-shad: 1px 1px #2a2a2e,
     accent: #0000ff33,
     accent-2: #9500ff,
     background-1: #2a2a2e,
@@ -22,5 +28,9 @@ $themes: (
     border-1: 2px solid #3a444c,
     underline: 1px solid white,
     mochi-guy: url("./assets/mochiSpin3.png"),
+    btn-grad: linear-gradient(225deg, #d6f76199 0%, #ffbc0599 100%),
+    btn-grad-hover: linear-gradient(160deg, rgba(168, 166, 23, 0.762) 0%, rgba(168, 166, 23, 0.76) 100%),
+    bg-opacity: 0.1,
+    bg-image-op: rgba(255,255,255, 0.1),
   ),
 );
